@@ -1,13 +1,15 @@
 import React from "react";
-import { BrowserRouter, Route } from "react-router-dom";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 import FormBuilder from "./form-builder";
 import FormEditor from "./components/form-editor/form-editor";
 
 const RootComponent: React.FC = () => {
   return (
     <BrowserRouter basename={`${window.spaBase}/form-builder`}>
-      <Route exact path="/" component={FormBuilder} />
-      <Route exact path="/edit/:uuid" component={FormEditor} />
+      <Routes>
+        <Route path="/" element={<FormBuilder />} />
+        <Route path="/edit/:uuid" element={<FormEditor />} />
+      </Routes>
     </BrowserRouter>
   );
 };
