@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
-import { TrashCan } from "@carbon/icons-react/next";
-import { Button, Column, Row } from "carbon-components-react";
+import { TrashCan } from "@carbon/react/icons";
+import { Button, Column, Row } from "@carbon/react";
 import { Section } from "../../../../api/types";
 import { SchemaContext } from "../../../../context/context";
 import EditSection from "../../modals/edit-section";
