@@ -12,9 +12,9 @@ import {
   SelectItem,
   TextInput,
   Toggle,
-} from "carbon-components-react";
+} from "@carbon/react";
 import { Section } from "../../../api/types";
-import { Edit } from "@carbon/icons-react/next";
+import { Edit } from "@carbon/react/icons";
 import { SchemaContext } from "../../../context/context";
 import { showToast } from "@openmrs/esm-framework";
 
